AsyncWebSocket: Fix packet buffer smash
In a misguided attempt to work around badly written clients,
AsyncWebSocketClient::_onData was trying to restore the value of the
byte one past the end of the packet buffer.  Unfortunately, that value
might have been *correctly* changed by the callback, such as when it's
storing heap metadata.

This seemed to be causing of websocket-related crashes in WLED on
ESP8266.