AsyncWebSocket: Use shared_ptr for buffers
Replace the hand-rolled buffer management logic with an RAII buffer
class based on std::shared_ptr.  This simplifies memory management and
seems to fix a memory leak, while also offering a clean move semantic
for clients.

The previous API is preserved through a compatibility object; it is
flagged as deprecated as the old API required a specialized allocator,
which is now unneccessary, so for compatibility reasons an extra heap
allocation is used.